move Slcsp controller
