+++ spec/controllers/individual_market/api/v1/slcsp_controller_spec.rb
@@ -2,16 +2,16 @@
 
 require 'rails_helper'
 
-describe Api::V1::SlcspController do
+describe IndividualMarket::Api::V1::SlcspController, skip: true do
 
   let(:request_xml) { File.read(Rails.root.join("spec", "test_data", "slcsp_payloads", "request.xml")) }
   let(:response_xml) { File.read(Rails.root.join("spec", "test_data", "slcsp_payloads", "response.xml")) }
   let(:plan) {double}
 
   context "valid request" do
-    xit 'returns https status 200' do
+    it 'returns https status 200' do
       allow(HappyMapper).to receive(:parse).with(anything).and_return(HappyMapper.parse(request_xml))
-      allow_any_instance_of(Api::V1::SlcspController).to receive(:find_slcsp).with(anything).and_return(plan)
+      allow_any_instance_of(IndividualMarket::Api::V1::SlcspController).to receive(:find_slcsp).with(anything).and_return(plan)
       allow_any_instance_of(ActionController::Rendering).to receive(:render).and_return(response_xml)
 
       post :plan, :format => "xml"
@@ -21,9 +21,9 @@ describe Api::V1::SlcspController do
   end
 
   context "invalid request" do
-    xit 'returns https status 422' do
+    it 'returns https status 422' do
       allow(HappyMapper).to receive(:parse).with(anything).and_raise(Exception.new)
-      allow_any_instance_of(Api::V1::SlcspController).to receive(:find_slcsp).with(anything).and_return(plan)
+      allow_any_instance_of(IndividualMarket::Api::V1::SlcspController).to receive(:find_slcsp).with(anything).and_return(plan)
 
       post :plan, params: { format: :xml }
       expect(response.status).to eq(422)
